+++ src/api/account/abstraction.ts
@@ -166,10 +166,10 @@ export class AccountAbstraction {
    */
   public isAccountAbstractionEnabled = async (args: {
     accountAddress: AccountAddressInput;
-    authenticationFunction: string;
+    authenticationFunction: MoveFunctionId;
   }) => {
     const functionInfos = await this.getAuthenticationFunction(args);
-    const { moduleAddress, moduleName, functionName } = getFunctionParts(args.authenticationFunction as MoveFunctionId);
+    const { moduleAddress, moduleName, functionName } = getFunctionParts(args.authenticationFunction);
     return (
       functionInfos?.some(
         (functionInfo) =>